+++ ProjectManager/ProjectManager/MainScene/View/ListView/CollectionView/ListCollectionView.swift
@@ -12,20 +12,20 @@ final class ListCollectionView: UICollectionView {
     private enum Section {
         case main
     }
-    private var viewModel: TodoListViewModel?
-    private var currentLongPressedCell: ListCell?
+    weak var transitionDelegate: TodoListViewControllerDelegate?
     private var todoDataSource: UICollectionViewDiffableDataSource<Section, Todo>?
     private var snapshot = NSDiffableDataSourceSnapshot<Section, Todo>()
     var category: String
+    var viewModel: TodoListViewModel
+    var currentLongPressedCell: ListCell?
     
     // MARK: Initializer
-    init(frame: CGRect, category: Category?, viewModel: TodoListViewModel?) {
-        super.init(frame: frame, collectionViewLayout: UICollectionViewLayout())
-        self.viewModel = viewModel
+    init(category: String, viewModel: TodoListViewModel) {
         self.category = category
-        delegate = self
+        self.viewModel = viewModel
+        super.init(frame: .zero, collectionViewLayout: UICollectionViewLayout())
         setupInitialView()
-        configureDataSource(with: viewModel?.read(category))
+        configureDataSource(with: viewModel.fetchTodoList(in: category))
         setupLongGestureRecognizerOnCollection()
     }
     
@@ -75,8 +75,7 @@ extension ListCollectionView {
                 completion(false)
                 return
             }
-            self.delete([item])
-            self.viewModel?.delete(model: item)
+            self.delete(todo: item)
             completion(true)
         }
         return UISwipeActionsConfiguration(actions: [deleteAction])
@@ -115,75 +114,5 @@ extension ListCollectionView {
         snapshot.appendSections([.main])
         snapshot.appendItems(items, toSection: .main)
         todoDataSource?.apply(snapshot, animatingDifferences: true)
-    }
-}
-
-// MARK: - CollectionView Delegate
-extension ListCollectionView: UICollectionViewDelegate {
-    
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        guard let model = viewModel?.read(category)?[indexPath.row] else { return }
-        viewModel?.goToEdit(model)
-    }
-}
-
-// MARK: - UIGestureRecognizerDelegate
-extension ListCollectionView: UIGestureRecognizerDelegate {
-    
-    private func setupLongGestureRecognizerOnCollection() {
-        let longPressedGesture = UILongPressGestureRecognizer(
-            target: self,
-            action: #selector(handleLongPress(gestureRecognizer:))
-        )
-        longPressedGesture.minimumPressDuration = 0.5
-        longPressedGesture.delegate = self
-        longPressedGesture.delaysTouchesBegan = true
-        addGestureRecognizer(longPressedGesture)
-    }
-    
-    @objc private func handleLongPress(gestureRecognizer: UILongPressGestureRecognizer) {
-        let location = gestureRecognizer.location(in: self)
-        let state = gestureRecognizer.state
-        switch state {
-        case .began:
-            animateLongPressBegin(at: location)
-        case .ended:
-            animateLongPressEnd(at: location)
-        default:
-            return
-        }
-    }
-    
-    private func animateLongPressBegin(at location: CGPoint) {
-        guard let indexPath = indexPathForItem(at: location) else { return }
-        UIView.animate(withDuration: 0.2) { [weak self] in
-            guard let self = self else { return }
-            guard let cell = self.cellForItem(at: indexPath) as? ListCell else { return }
-            self.currentLongPressedCell = cell
-            cell.transform = .init(scaleX: 0.95, y: 0.95)
-        }
-    }
-    
-    private func animateLongPressEnd(at location: CGPoint) {
-        UIView.animate(withDuration: 0.2) { [weak self] in
-            guard let self = self else { return }
-            guard let cell = self.currentLongPressedCell else { return }
-            cell.transform = .init(scaleX: 1, y: 1)
-            guard let indexPath = self.indexPathForItem(at: location) else { return }
-            guard cell == self.cellForItem(at: indexPath) as? ListCell else { return }
-            self.viewModel?.showPopover(
-                in: self,
-                location: (Double(location.x), Double(location.y)),
-                indexPath: indexPath.row
-            )
-        }
-    }
-}
-
-// MARK: - UIPopoverPresentationControllerDelegate
-extension ListCollectionView: UIPopoverPresentationControllerDelegate {
-    func adaptivePresentationStyle(for controller: UIPresentationController,
-                                   traitCollection: UITraitCollection) -> UIModalPresentationStyle {
-        return .none
     }
 }
